Init static variable correctly. Static variable should be avoided the most possible.
Fix problem that MSN and Yahoo weren't loaded at all.

CCMAIL: taupter@gmail.com


svn path=/branches/kopete/0.12/kopete/; revision=527378

+++ libkopete/avdevice/videodevicepool.h
@@ -122,7 +122,7 @@ protected:
 private:
 	VideoDevicePool();
 	static VideoDevicePool* s_self;
-	static __u64 m_clients; // Number of instances
+	static int m_clients; // Number of instances
 };
 
 }
